refactor: simpler compose
